[ADD] SingIn controller
