+++ src/main/java/com/hackathon/MYD/model/user/payload/request/SingInRequest.java
@@ -1,0 +1,9 @@
+package com.hackathon.MYD.model.user.payload.request;
+
+import lombok.Getter;
+
+@Getter
+public class SingInRequest {
+    private String nickname;
+    private String password;
+}
